Print out errors when ignoring them.
This is not ideal as a default behavior, but we are already printing
out the filename as we parse it, so it is consistent!  This makes it
easier for us to figure out why an IPDL file is not indexed in
searchfox by checking the index-log.  This could be better addressed
by introducing a concept of an error record that the parser could emit
into an analysis file so that we could display what's broken when
looking at the file in searchfox, potentially on an opt-in basis.