Add naive support for `[ManualAlloc, ChildImpl=virtual]` annotations.
The parser does not seem to like annotations inclusive of both the
`[ManualAlloc]` type added in
https://bugzilla.mozilla.org/show_bug.cgi?id=1736371 and the variations
that also include list of annotations in the patch that look like
`[ManualDealloc, ChildImpl=virtual, ParentImpl=virtual]`.

My efforts here are to try and get this to parse and technically
produce something usable, but I wouldn't be surprised if someone wanted
to properly consume the annotations that they would want something that
looks more strongly typed.  But this is an AST so maybe this is fine?